+++ src/main/java/com/github/jlangch/venice/impl/functions/ConcurrencyFunctions.java
@@ -2689,46 +2689,47 @@ public class ConcurrencyFunctions {
                 "preduce",
                 VncFunction
                     .meta()
-                    .arglists("(preduce n combine-fn combine-seed reduce-fn reduce-seed coll)")
+                    .arglists(
+                    	"(preduce n combine-fn combine-seed reduce-fn reduce-seed coll)",
+                    	"(preduce n reduce-fn reduce-seed coll)")
                     .doc(
                         "Reduces a collection using a parallel reduce-combine strategy. " +
                         "The collection is partitioned into groups of approximately " +
                         "n items, each of which is reduced with reduce-fn (with reduce-seed " +
                         "as its seed value). The results of these reductions are then reduced " +
-                        "with combine-fn (with combine-seed as its seed value).")
+                        "with combine-fn (with combine-seed as its seed value).\n" +
+                        "Withhout an explicit combine-fn the reduce-fn and its seed reduce-seed " +
+                        "will be used as combine-fn and combine-seed.")
                     .examples(
                         "(preduce 3 + 0 + 0 [1 2 3 4 5])",
-                        "(preduce 3 (fn [acc x] (+ acc x)) 0 (fn [acc x] (+ acc x)) 0 [1 2 3 4 5])")
+                        "(preduce 3 (fn [acc x] (+ acc x)) 0 (fn [acc x] (+ acc x)) 0 [1 2 3 4 5])",
+                        "(preduce 3 + 0 [1 2 3 4 5])",
+                        "(preduce 3 (fn [acc x] (+ acc x)) 0 [1 2 3 4 5])")
                     .seeAlso("reduce", "map", "filter", "pmap", "pcalls")
                     .build()
         ) {
             @Override
             public VncVal apply(final VncList args) {
-                ArityExceptions.assertArity(this, args, 6);
+                ArityExceptions.assertArity(this, args, 4, 6);
 
-                final VncVal n = Coerce.toVncLong(args.nth(0));
-                final VncVal combineFn = args.nth(1);
-                final VncVal combineSeed = args.nth(2);
-                final VncVal reduceFn = args.nth(3);
-                final VncVal reduceSeed = args.nth(4);
-                final VncSequence seq = Coerce.toVncSequence(args.nth(5));
+                if (args.size() == 4) {
+                    final VncVal n = Coerce.toVncLong(args.nth(0));
+                    final VncVal reduceFn = args.nth(1);
+                    final VncVal reduceSeed = args.nth(2);
+                    final VncSequence seq = Coerce.toVncSequence(args.nth(3));
 
-                // reducing (parallel)
-                final VncList chunks = (VncList)CoreFunctions.partition_all.applyOf(n, seq);
-                final List<VncVal> jobs = new ArrayList<>();
-                chunks.forEach(c ->
-                	jobs.add(
-                		future.applyOf(
-                			VncFunction.of(
-                				() -> CoreFunctions.reduce.applyOf(reduceFn, reduceSeed, c)))));
-
-                VncList reductions = VncList.empty();
-                for(VncVal job : jobs) {
-                    reductions = reductions.addAtEnd(CoreConcurrencyFunctions.deref.applyOf(job));
+                    return preduce(n, reduceFn, reduceSeed, reduceFn, reduceSeed, seq);
                 }
+                else {
+                    final VncVal n = Coerce.toVncLong(args.nth(0));
+                    final VncVal combineFn = args.nth(1);
+                    final VncVal combineSeed = args.nth(2);
+                    final VncVal reduceFn = args.nth(3);
+                    final VncVal reduceSeed = args.nth(4);
+                    final VncSequence seq = Coerce.toVncSequence(args.nth(5));
 
-                // combining (sequential)
-                return CoreFunctions.reduce.applyOf(combineFn, combineSeed, reductions);
+                    return preduce(n, combineFn, combineSeed, reduceFn, reduceSeed, seq);
+                }
             }
 
             private static final long serialVersionUID = -1848883965231344442L;
@@ -2749,6 +2750,31 @@ public class ConcurrencyFunctions {
         mngdExecutor.setMaximumThreadPoolSize(maximumPoolSize);
     }
 
+    private static VncVal preduce(
+            final VncVal n,
+            final VncVal combineFn,
+            final VncVal combineSeed,
+            final VncVal reduceFn,
+            final VncVal reduceSeed,
+            final VncSequence seq
+    ) {
+        // reducing (parallel)
+        final VncList chunks = (VncList)CoreFunctions.partition_all.applyOf(n, seq);
+        final List<VncVal> jobs = new ArrayList<>();
+        chunks.forEach(c ->
+        	jobs.add(
+        		future.applyOf(
+        			VncFunction.of(
+        				() -> CoreFunctions.reduce.applyOf(reduceFn, reduceSeed, c)))));
+
+        VncList reductions = VncList.empty();
+        for(VncVal job : jobs) {
+            reductions = reductions.addAtEnd(CoreConcurrencyFunctions.deref.applyOf(job));
+        }
+
+        // combining (sequential)
+        return CoreFunctions.reduce.applyOf(combineFn, combineSeed, reductions);
+    }
 
     private static void safelyCancelFuture(final Future<?> future) {
         try {