+++ ChangeLog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 
 
-## [1.10.33] - 2023-01-xx
+## [1.10.33] - 2023-01-18
 
 ### Added